fix(lens): re-extract switch-in center-on-active into useCanvasCenterOnSwitch
The pendingCenter signal + two effects + the activeScope camera reach-in move out of TerminalCanvas into canvas/useCanvasCenterOnSwitch.ts, taking activeTileLayout as its one param and mounted like its sibling hooks; behavior is a byte-identical relocation.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-4, raised by lowy). Not pushed or merged.
