+++ packages/client/src/canvas/TerminalCanvas.tsx
@@ -33,15 +33,13 @@ import {
   Show,
   Switch,
 } from "solid-js";
-import { activeScope } from "../hostScope/hostScopes";
 import { useStaleCheck } from "../terminal/staleness";
 import { useTerminalStore } from "../terminal/useTerminalStore";
 import type { TileId } from "../tile/tileContent";
 import { useTileStore } from "../tile/useTileStore";
-import { activeHost, savedSessionSub } from "../wire";
+import { savedSessionSub } from "../wire";
 import CanvasMinimap from "./CanvasMinimap";
 import CanvasTile, { type CanvasTileMode } from "./CanvasTile";
-import { switchInNeedsCenter } from "./cameraSwap";
 import Dock from "./dock/Dock";
 import { applyResize, type ResizeDirection } from "./resizeGeometry";
 import type { TileLayout } from "./TileLayout";
@@ -51,6 +49,7 @@ import {
   findFreeTilePosition,
 } from "./tilePlacement";
 import { planTilePlacements } from "./tilePlacementPlan";
+import { useCanvasCenterOnSwitch } from "./useCanvasCenterOnSwitch";
 import { useCanvasFocus } from "./useCanvasFocus";
 import { usePendingLayouts } from "./usePendingLayouts";
 import { useTileAura } from "./useTileAura";
@@ -151,47 +150,16 @@ const TerminalCanvas: Component<{
     return result;
   });
 
-  // Per-host camera center-on-active on host switch. The camera pose is now
-  // per-host and RETAINED in `hostScopes.active().camera`, so a switch shows that
-  // host's saved pose BY CONSTRUCTION — there is no snapshot/restore step to race
-  // the incoming tile's mount (the class the deleted `useCanvasCameraSwap` bridge
-  // could never close). What remains is the switch-in center DECISION, deferred
-  // past the mount race: once the incoming host's active tile has a measured
-  // layout, seed a never-positioned host on it (firstVisit) or re-center a host
-  // whose active tile drifted out of its retained view (stale) — the pure
-  // `switchInNeedsCenter` core and the layout-gated `focus.request` pan unchanged.
+  // Per-host camera center-on-active on host switch. The decision (seed a
+  // never-positioned host on its active tile, re-center a host whose tile drifted
+  // out of its retained view) lives in `useCanvasCenterOnSwitch`, mounted here
+  // like the sibling hooks above; its one canvas-derived input is the active
+  // tile's resolved layout over the pending-layout overrides.
   const activeTileLayout = (): TileLayout | undefined => {
     const id = tileStore.activeId();
     return id ? layoutOf(id) : undefined;
   };
-  const [pendingCenter, setPendingCenter] = createSignal(false);
-  createEffect(
-    on(activeHost, (_curr, prev) => {
-      if (prev === undefined) return; // initial mount — no switch to service yet
-      setPendingCenter(true);
-    }),
-  );
-  createEffect(() => {
-    if (!pendingCenter()) return;
-    const activeTile = activeTileLayout();
-    if (!activeTile) return; // mount race — wait for the tile to be measured
-    const camera = activeScope()?.camera;
-    if (!camera) {
-      setPendingCenter(false);
-      return;
-    }
-    const { width, height } = viewport.viewportSize();
-    if (
-      switchInNeedsCenter(
-        camera.positioned() ? camera.snapshot() : null,
-        activeTile,
-        width,
-        height,
-      )
-    )
-      store.requestCenterActive();
-    setPendingCenter(false); // consume once resolved
-  });
+  useCanvasCenterOnSwitch(activeTileLayout);
 
   // Auto-assign a default layout for tiles with no saved position. A new
   // tile opens at the viewport-center cascade and NOTHING ELSE MOVES —
